# apps/backend/internal/problem/delivery/http/http.go
package http

import (
	"context"
	"fmt"
	"strconv"

	"problum/internal/api"
	"problum/internal/config"
	"problum/internal/problem/service"
	"problum/internal/problem/service/dto"

	"github.com/gofiber/fiber/v3"
	"github.com/rs/zerolog/log"
)

//go:generate go run go.uber.org/mock/mockgen -source=http.go -destination=mocks/mock_service.go -package=mocks
type Service interface {
	GetWithOptions(context.Context, int, ...service.Option) (*dto.Problem, error)
	Submit(context.Context, *dto.ProblemSubmit) (int, error)
}

type Handler struct {
	cfg *config.Config
	svc Service
}

func New(cfg *config.Config, svc Service) *Handler {
	return &Handler{
		cfg: cfg,
		svc: svc,
	}
}

// Get returns a problem.
// @Summary Get problem
// @Description Returns a problem by ID. Pass language to include a starter template for that language.
// @Tags problems
// @Produce json
// @Param courseID path int true "Course ID"
// @Param problemID path int true "Problem ID"
// @Param language query string false "Template language"
// @Success 200 {object} api.ProblemGetResponse
// @Failure 401 {string} string
// @Failure 403 {string} string
// @Failure 500 {string} string
// @Security BearerAuth
// @Router /courses/{courseID}/problems/{problemID} [get]
func (h *Handler) Get(c fiber.Ctx) error {
	id, err := strconv.Atoi(c.Params("problemID"))
	if err != nil {
		log.Error().Err(err).Msg("Failed to parse problem id in query param")
		return fmt.Errorf("failed to parse problem id in query param")
	}

	language := c.Query("language")

	problem, err := h.svc.GetWithOptions(
		c.Context(),
		id,
		service.WithTemplate(),
		service.WithLanguage(language),
	)
	if err != nil {
		log.Error().Err(err).Msg("Failed to get problem")
		return fmt.Errorf("failed to get problem")
	}

	return c.JSON(dto.ToAPI(problem))
}

// Submit submits a solution.
// @Summary Submit solution
// @Description Submits code for a problem and returns the created attempt ID.
// @Tags problems
// @Accept json
// @Produce json
// @Param courseID path int true "Course ID"
// @Param problemID path int true "Problem ID"
// @Param request body api.ProblemSubmitRequest true "Problem submit request"
// @Success 200 {object} api.ProblemSubmitResponse
// @Failure 401 {string} string
// @Failure 403 {string} string
// @Failure 500 {string} string
// @Security BearerAuth
// @Router /courses/{courseID}/problems/{problemID}/submit [post]
func (h *Handler) Submit(c fiber.Ctx) error {
	submitReq := &api.ProblemSubmitRequest{}
	if err := c.Bind().JSON(submitReq); err != nil {
		return err
	}

	userID, ok := c.Locals("user_id").(int)
	if !ok {
		log.Error().Msg("Failed to get userID")
		return fmt.Errorf("Failed to get userID")
	}

	problemID, err := strconv.Atoi(c.Params("problemID"))
	if err != nil {
		log.Error().Err(err).Msg("Failed to parse problem id in query param")
		return fmt.Errorf("failed to parse problem id in query param")
	}

	attemptID, err := h.svc.Submit(c.Context(), &dto.ProblemSubmit{
		ProblemID: problemID,
		UserID:    userID,
		Language:  submitReq.Language,
		Code:      submitReq.Code,
	})
	if err != nil {
		log.Error().Err(err).Msg("Failed to submit")
		return fmt.Errorf("failed to submit")
	}

	return c.JSON(api.ProblemSubmitResponse{
		AttemptID: attemptID,
	})
}